Implemented refactor to redistributing files in api folder
