+++ app/src/main/java/com/fit2081/arrtish/id32896786/a1/api/fruityvice/FruityViceDataClasses.kt
@@ -1,4 +1,4 @@
-package com.fit2081.arrtish.id32896786.a1.nutricoach
+package com.fit2081.arrtish.id32896786.a1.api.fruityvice
 
 data class Fruit(
     val genus: String,
