+++ app/src/main/java/com/fit2081/arrtish/id32896786/a1/api/gpt/ChatGptApi.kt
@@ -1,8 +1,7 @@
-package com.fit2081.arrtish.id32896786.a1.gpt
+package com.fit2081.arrtish.id32896786.a1.api.gpt
 
 
 import retrofit2.http.Body
-import retrofit2.http.Headers
 import retrofit2.http.POST
 
 interface ChatGptApi {
